+++ NEWS.md
@@ -1,3 +1,8 @@
+# ClinicalCharacteristics 1.1.0
+
+* Add totSubjects column to clarify denominator
+* Add function for observedTimeToFirst
+
 # ClinicalCharacteristics 1.0.0
 
 * Re-structure ClinChar using R6 classes